+++ src/dbtarzan/db/actor/CopyWorker.scala
@@ -9,6 +9,7 @@ import dbtarzan.config.connections.ConnectionData
 import dbtarzan.config.password.EncryptionKey
 import dbtarzan.db.util.ResourceManagement.using
 import dbtarzan.db._
+import dbtarzan.db.foreignkeys.{ ForeignKeyLoader, ForeignKeysToFile }
 import dbtarzan.messages._
 import dbtarzan.localization.Localization
